pyproject.toml
[tool.poetry]
name = "iris-classifier"
version = "0.1.0"
description = ""
authors = ["danil-tr <daniltrvit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.3"
pandas = "^2.2.0"
joblib = "^1.3.2"
scikit-learn = "^1.4.0"
dvc = "^3.42.0"
pyarrow = "^15.0.0"
dvc-gdrive = "^3.0.1"
hydra-core = "^1.3.2"
boto3 = "^1.34.34"
mlflow = "^2.10.0"
catboost = "^1.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pre-commit = "^3.6.0"

[tool.black]
line-length = 90
target-version = ["py39"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
